+++ src/SharedKernel/Domain/OptionDto.cs
@@ -1,3 +1,3 @@
 namespace SharedKernel.Domain;
 
-public sealed record OptionDto(string Label,int Value,string Description,Guid? AdditionalInfo=default,bool IsTaken=false);
+public sealed record OptionDto(string Label,long Value,string Description,Guid? AdditionalInfo=default,bool IsTaken=false);
